requirements.txt
Flask==2.3.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
authlib==1.2.1
requests-oauthlib==1.3.1
openai==1.3.0